Moves tests to root of project
