+++ tests/typescript/create-component-store.tsx
