+++ tests/typescript/create-store.ts
