+++ tests/typescript/generic-model.ts
