+++ tests/typescript/issue117.ts
