+++ tests/typescript/issue224.ts
