+++ tests/typescript/issue246.ts
