+++ tests/typescript/issue66.tsx
