+++ tests/typescript/issue87.ts
