+++ tests/typescript/listener-thunk.ts
