+++ tests/typescript/react-redux-compat.tsx
